BB-7117: Payment term relations with customer and customer group are missed on update from early versions of commerce

+++ src/Oro/Bundle/PaymentTermBundle/Migrations/Schema/OroPaymentTermBundleInstaller.php
@@ -126,7 +126,7 @@ class OroPaymentTermBundleInstaller implements
         $this->paymentTermExtension->addPaymentTermAssociation($schema, 'oro_customer');
         $this->paymentTermExtension->addPaymentTermAssociation($schema, 'oro_customer_group');
         
-        $this->migrateRelations($queries);
+        $this->migrateRelations($schema, $queries);
 
         $associationTableName = $this->activityExtension->getAssociationTableName('oro_note', self::TABLE_NAME);
         if (!$schema->hasTable($associationTableName)) {
@@ -146,9 +146,10 @@ class OroPaymentTermBundleInstaller implements
     }
 
     /**
+     * @param Schema $schema
      * @param QueryBag $queries
      */
-    protected function migrateRelations(QueryBag $queries)
+    protected function migrateRelations(Schema $schema, QueryBag $queries)
     {
         if ($this->platform instanceof MySqlPlatform) {
             $queryAccount = <<<QUERY
@@ -186,10 +187,15 @@ QUERY;
             new RemoveFieldQuery('Oro\Bundle\PaymentTermBundle\Entity\PaymentTerm', 'accountGroups')
         );
 
-        $queries->addPostQuery($queryAccount);
-        $queries->addPostQuery($queryGroup);
-        $queries->addPostQuery('DROP TABLE oro_payment_term_to_account;');
-        $queries->addPostQuery('DROP TABLE oro_payment_term_to_acc_grp;');
+        if ($schema->hasTable('oro_payment_term_to_account')) {
+            $queries->addPostQuery($queryAccount);
+            $queries->addPostQuery('DROP TABLE oro_payment_term_to_account;');
+        }
+
+        if ($schema->hasTable('oro_payment_term_to_acc_grp')) {
+            $queries->addPostQuery($queryGroup);
+            $queries->addPostQuery('DROP TABLE oro_payment_term_to_acc_grp;');
+        }
     }
 
     /**
